Created a class for storing world objects
